+++ pika/src/pika_binlog_sender_thread.cc
@@ -141,6 +141,7 @@ unsigned int PikaBinlogSenderThread::ReadPhysicalRecord(slash::Slice *result) {
     length = 1;
     type = kBadRecord;
   }
+
   //std::cout<<"2 --> con_offset_: "<<con_offset_<<" last_record_offset_: "<<last_record_offset_<<std::endl;
   s = queue_->Read(length, &buffer_, backing_store_);
   *result = slash::Slice(buffer_.data(), buffer_.size());