+++ pika/third/rocksdb/utilities/write_batch_with_index/write_batch_with_index_internal.cc
@@ -71,13 +71,16 @@ Status ReadableWriteBatch::GetEntryFromDataOffset(size_t data_offset,
       break;
     case kTypeNoop:
     case kTypeBeginPrepareXID:
+    case kTypeBeginPersistedPrepareXID:
+    case kTypeBeginUnprepareXID:
     case kTypeEndPrepareXID:
     case kTypeCommitXID:
     case kTypeRollbackXID:
       *type = kXIDRecord;
       break;
     default:
-      return Status::Corruption("unknown WriteBatch tag");
+      return Status::Corruption("unknown WriteBatch tag ",
+                                ToString(static_cast<unsigned int>(tag)));
   }
   return Status::OK();
 }